backend/requirements.txt
alembic==1.16.1
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib>=1.2.0
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.2.6
pandas==2.2.3
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
SQLAlchemy==2.0.41
threadpoolctl==3.6.0
typing_extensions==4.13.2
tzdata==2025.2
Werkzeug==3.1.3
